fix all bugs
