+++ src/components/article/Article.tsx
@@ -1,11 +1,10 @@
-import clsx from 'clsx';
 import styles from './Article.module.scss';
 import plane from 'src/images/plane.png';
 import { Text } from 'src/ui/text';
 
 export const Article = () => {
 	return (
-		<article className={clsx(styles.article)}>
+		<article className={styles.article}>
 			<Text as='h1' size={45} weight={800} uppercase dynamicLite>
 				Портрет Западной Швейцарии
 			</Text>
